convert types to decimals

+++ src/settings.rs
@@ -1,11 +1,12 @@
 use config::{Config, ConfigError, Environment};
 use kafka_settings::KafkaSettings;
+use rust_decimal::Decimal;
 use serde::Deserialize;
 
 #[derive(Debug, Deserialize)]
 pub struct AppSettings {
-    pub initial_equity: f64,
-    pub internal_leverage: f64,
+    pub initial_equity: Decimal,
+    pub internal_leverage: Decimal,
     pub batch_seconds: u64,
 }
